add remaining flows
