+++ pipelines/bases/br_ms_vacinacao_covid19/__init__.py
@@ -1,0 +1,7 @@
+"""
+Prefect flows for br_ibge_ipca project
+"""
+###############################################################################
+# Automatically managed, please do not touch
+###############################################################################
+from .flows import *
